covid19-data-analysis/requirements.txt
# Data Processing and Analysis
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine Learning
scikit-learn==1.3.0
statsmodels==0.14.0

# Web Dashboard
streamlit==1.28.0
dash==2.14.1

# Jupyter Notebooks
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.25.0

# Configuration Management
PyYAML==6.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
tqdm==4.65.0
openpyxl==3.1.2